Fix cannot find endpoint

+++ BeatTogether.MasterServer.Kernel/Implementations/DedicatedServerEventHandler.cs
@@ -69,7 +69,8 @@ namespace BeatTogether.MasterServer.Kernel.Implementations
 
         private Task HandlePlayerDisconnect(PlayerLeaveServerEvent integrationEvent)
         {
-            MasterServerSession session = _masterServerSessionService.GetSession((EndPoint)IPEndPoint.Parse(integrationEvent.endPoint));
+            if (!_masterServerSessionService.TryGetSession((EndPoint)IPEndPoint.Parse(integrationEvent.endPoint), out var session))
+                return Task.CompletedTask;
             _serverRepository.DecrementCurrentPlayerCount(session.Secret);
             _masterServerSessionService.RemoveSecretFromSession(session.EndPoint);
             return Task.CompletedTask;
